Move some ncomponents into the appropriate game-code modules [Re: nmkataoka/adv-life#110]
