+++ src/1-game-code/Combat/StatusEffectConstants.ts
