+++ src/1-game-code/Combat/StatusEffectsSys.ts
@@ -1,6 +1,6 @@
 import { ECSystem } from '../../0-engine/ECS/ECSystem';
 import { GetComponentManager } from '../../0-engine';
-import { StatusEffectsCmpt } from '../ncomponents/StatusEffectsCmpt';
+import { StatusEffectsCmpt } from './StatusEffectsCmpt';
 
 export class StatusEffectsSys extends ECSystem {
   public Start(): void {}
