+++ src/1-game-code/Items/WearableCmpt.ts
@@ -1,4 +1,4 @@
-import { NComponent } from '../../../0-engine';
+import { NComponent } from '../../0-engine';
 
 /* An item is worn if it won't fall off if the user falls unconscious */
 export class WearableCmpt implements NComponent {
