+++ src/1-game-code/Items/WieldableCmpt.ts
@@ -1,4 +1,4 @@
-import { NComponent } from '../../../0-engine';
+import { NComponent } from '../../0-engine';
 
 export class WieldableCmpt implements NComponent {
   public damage = 15;
