+++ src/1-game-code/Town/TownLocationsCmpt.ts
@@ -1,6 +1,5 @@
-import { NComponent } from '../../../0-engine';
-import { IEntityRelationship } from '../../Agent/ConditionSet/EntityRelationshipTemplate';
-import { Entity } from '../../../0-engine';
+import { NComponent, Entity } from '../../0-engine';
+import { IEntityRelationship } from '../Agent/ConditionSet/EntityRelationshipTemplate';
 
 export class TownLocationsCmpt implements NComponent, IEntityRelationship {
   public locationIds: Entity[] = [];
