+++ src/1-game-code/Town/index.ts
@@ -1,0 +1,1 @@
+export { TownLocationsCmpt } from './TownLocationsCmpt';
