--- src/1-game-code/ncomponents/Town/index.ts
@@ -1,1 +1,0 @@
-export * from './TownLocationsCmpt';
