+++ src/1-game-code/ncomponents/index.ts
@@ -1,16 +1,14 @@
 export { AgentCmpt } from './AgentCmpt';
 export { CanAttackCmpt } from './CanAttackCmpt';
 export { ClassCmpt } from './ClassCmpt';
-export { CombatPositionCmpt } from './CombatPositionCmpt';
-export { CombatStatsCmpt } from './CombatStatsCmpt';
+export { CombatPositionCmpt } from '../Combat/CombatPositionCmpt';
+export { CombatStatsCmpt } from '../Combat/CombatStatsCmpt';
 export { FactionCmpt } from './FactionCmpt';
 export { HealthCmpt } from './HealthCmpt';
 export { HeldItemsCmpt, InventoryCmpt, WornItemsCmpt } from '../Inventory';
-export * from './Items';
+export * from '../Items';
 export { NameCmpt } from '../../0-engine/ECS/built-in-components/NameCmpt';
 export type { PersonalityArray } from './PersonalityCmpt';
 export { PersonalityCmpt } from './PersonalityCmpt';
 export { PlayerCmpt } from './PlayerCmpt';
 export { RaceCmpt } from './RaceCmpt';
-export { StatusEffectsCmpt } from './StatusEffectsCmpt';
-export * from './Town';
