+++ src/3-frontend-api/SkillData.ts
@@ -1,5 +1,5 @@
 import { EntityManager } from '../0-engine';
-import { CombatStatsCmpt } from '../1-game-code/ncomponents/CombatStatsCmpt';
+import { CombatStatsCmpt } from '../1-game-code/Combat/CombatStatsCmpt';
 
 export function getSkillData(user: number, targets: number[], skillName: string) {
   let data: any;
