+++ src/3-frontend-api/UnitInfo.ts
@@ -1,9 +1,9 @@
 import { GameManager } from '../0-engine/GameManager';
-import { CombatStatsCmpt } from '../1-game-code/ncomponents/CombatStatsCmpt';
+import { CombatStatsCmpt } from '../1-game-code/Combat/CombatStatsCmpt';
 import { HealthCmpt } from '../1-game-code/ncomponents/HealthCmpt';
 import { FactionCmpt } from '../1-game-code/ncomponents/FactionCmpt';
-import { CombatPositionCmpt } from '../1-game-code/ncomponents/CombatPositionCmpt';
-import { StatusEffectsCmpt } from '../1-game-code/ncomponents/StatusEffectsCmpt';
+import { CombatPositionCmpt } from '../1-game-code/Combat/CombatPositionCmpt';
+import { StatusEffectsCmpt } from '../1-game-code/Combat/StatusEffectsCmpt';
 import { GetView } from '../0-engine';
 import { GoalQueueCmpt } from '../1-game-code/Agent/GoalQueueCmpt';
