+++ src/3-frontend-api/town/getTowns.ts
@@ -1,5 +1,5 @@
 import { GetView } from '../../0-engine';
-import { TownLocationsCmpt } from '../../1-game-code/ncomponents';
+import { TownLocationsCmpt } from '../../1-game-code/Town';
 import { TownsDict } from './TownInfo';
 import { GameManager } from '../../0-engine/GameManager';
 import { getNameCmpt } from '../getName';
